Shut it Ashley
Kinda works but also fuck you, I hate this job. I just wanted to make a nice dark theme for the application and I got landed with this shit. I wanna die.  Fuck you and goodbye

+++ public/src/app/components/quote-box/quote-box.component.ts
@@ -1,6 +1,9 @@
 import { Component, OnInit, Input } from '@angular/core';
 import Quote from 'src/app/models/quote';
 import { QuoteService } from 'src/app/services/quote.service';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
+import { ModalComponent } from 'src/app/components/add_quote_modal/modal.component';
+import { FormControl } from '@angular/forms'
 
 
 @Component({
@@ -10,10 +13,13 @@ import { QuoteService } from 'src/app/services/quote.service';
 })
 export class QuoteBoxComponent implements OnInit {
 
-  @Input() 
+  @Input()
   quote: Quote;
+  quoteBody = new FormControl();
+  quotee = new FormControl();
+  dateQuotedAsString = new FormControl();
 
-  constructor(private quoteService: QuoteService) { }
+  constructor(private quoteService: QuoteService, private modalService: NgbModal) { }
 
   ngOnInit(): void {
   }
@@ -24,14 +30,46 @@ export class QuoteBoxComponent implements OnInit {
   }
 
   deleteQuote() {
-    this.quoteService.deleteQuote(this.quote._id).subscribe(
-      (data) => {
-        window.location.reload();
-      },
-      (err) => {
-        
-      }
-    );
+    if(confirm("Are you sure?")) {
+      this.quoteService.deleteQuote(this.quote._id).subscribe(
+        (data) => {
+          window.location.reload();
+        },
+        (err) => {
+
+        }
+      );
+    }
   }
 
-} 
+  openModal(content) {
+    this.quoteBody = new FormControl(this.quote.body);
+    this.quotee = new FormControl(this.quote.quotee);
+    this.dateQuotedAsString = new FormControl(this.dateToString());
+    this.modalService.open(content, { size: 'lg' });
+  }
+  patchQuote() {
+    let dateQuoted = this.getDateQuotedAsISO();
+    let quoteBody = this.quoteBody.value;
+    let quotee = this.quotee.value;
+
+    this.quoteService.patchQuote(this.quote._id, quotee, quoteBody, dateQuoted).subscribe(
+      data => {
+        window.location.reload();
+      },
+      err => {
+        console.log(err);
+      }
+    )
+
+  }
+  private getDateQuotedAsISO() {
+    let date = this.dateQuotedAsString.value;
+    return new Date(date.year, date.month - 1, date.day).toISOString();
+  }
+  private dateToString() {
+    let date = new Date(this.quote.dateQuoted);
+    return `${date.getDate().toString()}-${(date.getMonth() + 1).toString()}-${date.getFullYear().toString()}`;
+  }
+
+}
